config: the model-specific declarations leave overrides.py
The 2026-07-03 placement ruling said the config-time declarations stay in
arg_groups/ -- the model modules are not imported at __post_init__, and
the frontend process must not depend on a GPU kernel import -- but that
they could split out mechanically once the families grew. 27 handlers and
992 lines is grown: overrides.py goes 3043 -> 1931, and each family is
now one module under model_overrides/, mirroring the models/ naming.

Three things this turned up that were not mechanical:

- __init__.py is a list of imports, and importing is what registers, so
  that list decides which declaration wins for the six architectures two
  families both claim. isort alphabetised it and flipped InternS2Mobius.
  The file is marked `# isort: skip_file`, and test_model_override_split
  pins the order so the next tool to reorder it fails loudly.

- The registries are read off model_override_base rather than through
  imported copies of the names. The registrars append there; a copied
  name is a second binding, and the test fixture that isolated the
  registry was rebinding one while the registrar wrote to the other.

- Each family module logs under its own name, so a warning now says which
  family raised it instead of a single `arg_groups.overrides`.

The chain ratchet scanned overrides.py by path for the declaration bodies
and *silently skipped* what it could not find, which is the vacuous
census its own docstring warns about. It now resolves each callable's
source from its module and asserts it found one.

Registry compares identical (55 architectures, 61 registrations) and the
62-shape probe is unchanged.
