+++ python/sglang/srt/arg_groups/model_overrides/__init__.py
@@ -1,0 +1,42 @@
+"""Per-model config-time override declarations, one module per family,
+mirroring the ``models/`` naming.
+
+Importing this package is what registers them, so **this file is the order**.
+It is the order the declarations had in ``overrides.py``, and it is
+load-bearing: some architectures are claimed by more than one family module
+(``InternS2MobiusForConditionalGeneration`` by both ``qwen3_5`` and
+``interns2_mobius``), and the gate applies matching declarations in
+registration order, last writer winning. Alphabetising this list silently
+changes which declaration wins -- which is why isort is switched off here, and
+why ``test_model_override_split.py`` pins the order.
+"""
+
+# isort: skip_file
+# fmt: off
+from sglang.srt.arg_groups.model_overrides import kimi_k3  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import deepseek_v2  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import mimo_v2  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import minimax_m2  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import minimax_m3  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import gemma2_gemma3  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import exaone  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import gpt_oss  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import llama4  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import gemma4  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import moss_vl  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import minicpm  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import minicpmv  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import falcon_h1  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import granitemoehybrid  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import lfm2  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import deepseek_v4  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import inkling  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import nemotron_h  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import qwen3_5  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import interns2_mobius  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import qwen3_vl  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import qwen3_moe  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import glm4_moe  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import olmo2  # noqa: F401
+from sglang.srt.arg_groups.model_overrides import muse_glimmer  # noqa: F401
+# fmt: on
